Change: Make the RailDriver work with both x32 and x64 builds
WINDOWS ONLY
